pyproject.toml
[tool.poetry]
name = "OsmoseApp"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2"
djangorestframework = "^3.12.4"
djangorestframework-simplejwt = "^4.6.0"
gunicorn = "^20.1.0"
psycopg2-binary = "^2.8.6"
sentry-sdk = "^1.3.1"
django-tinymce = "^3.6.1"
python-dateutil = ">=2.4"
metadatax = { path = "metadatax-0.2.2.tar.gz" }
drf-yasg = {extras = ["validation"], version = "^1.21.7"}


[tool.poetry.group.dev.dependencies]
pytest = "^5.2"
coverage = { extras = ["toml"], version = "^6.3" }
black = "^22.10.0"
pylint = "2.14.5"
pylint-django = "^2.5.3"
django-debug-toolbar = "^4.0.0"
faker = "^19.3.1"
freezegun = "^1.2.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.coverage.run]
source = ["backend"]
branch = true
omit = ["backend/asgi.py", "backend/wsgi.py", "backend/settings/*", "backend/api/migrations/*", "**/__init__.py"]

[tool.coverage.report]
show_missing = true

[tool.pylint.messages_control]
disable = ["too-many-ancestors", ]


[tool.pylint.MASTER]
load-plugins = "pylint_django"
django-settings-module = "backend.settings"

[tool.pylint.MAIN]
ignore-paths = ["backend\\api\\migrations|backend/api/migrations", "backend\\osmosewebsite\\migrations|backend/osmosewebsite/migrations"]

[tool.pylint.FORMAT]
max-line-length = 120
good-names = ["pk"]

[tool.pylint.DESIGN]
max-locals = 20
